silm down the movement gamestream

+++ bomberman.js
@@ -36,18 +36,18 @@ function create() {
     game.physics.startSystem(Phaser.Physics.ARCADE);
     game.physics.arcade.setBoundsToWorld();
 
-        game.input.keyboard.addKey(Phaser.Keyboard.LEFT).onDown.add(function(){gameStream.emit('jimmyLeft',  {key: true})});
-        game.input.keyboard.addKey(Phaser.Keyboard.RIGHT).onDown.add(function(){gameStream.emit('jimmyRight',  {key: true})});
-        game.input.keyboard.addKey(Phaser.Keyboard.UP).onDown.add(function(){gameStream.emit('jimmyUp',  {key: true})});
-        game.input.keyboard.addKey(Phaser.Keyboard.DOWN).onDown.add(function(){gameStream.emit('jimmyDown',  {key: true})});
-        game.input.keyboard.addKey(Phaser.Keyboard.LEFT).onUp.add(function(){gameStream.emit('jimmyLeft',  {x: jimmy.x, y: jimmy.y, key: false})});
-        game.input.keyboard.addKey(Phaser.Keyboard.RIGHT).onUp.add(function(){gameStream.emit('jimmyRight',  {x: jimmy.x, y: jimmy.y, key: false})});
-        game.input.keyboard.addKey(Phaser.Keyboard.UP).onUp.add(function(){gameStream.emit('jimmyUp',  {x: jimmy.x, y: jimmy.y, key: false})});
-        game.input.keyboard.addKey(Phaser.Keyboard.DOWN).onUp.add(function(){gameStream.emit('jimmyDown',  {x: jimmy.x, y: jimmy.y, key: false})});
+        game.input.keyboard.addKey(Phaser.Keyboard.LEFT).onDown.add(function(){gameStream.emit('Movement',  {dir:"left", name: jimmy.name, key: true})});
+        game.input.keyboard.addKey(Phaser.Keyboard.RIGHT).onDown.add(function(){gameStream.emit('Movement',  {dir:"right", name: jimmy.name, key: true})});
+        game.input.keyboard.addKey(Phaser.Keyboard.UP).onDown.add(function(){gameStream.emit('Movement',  {dir:"up", name: jimmy.name, key: true})});
+        game.input.keyboard.addKey(Phaser.Keyboard.DOWN).onDown.add(function(){gameStream.emit('Movement',  {dir:"down", name: jimmy.name, key: true})});
+        game.input.keyboard.addKey(Phaser.Keyboard.LEFT).onUp.add(function(){gameStream.emit('Movement',  {x: jimmy.x, y: jimmy.y, dir:"left", name: jimmy.name, key: false})});
+        game.input.keyboard.addKey(Phaser.Keyboard.RIGHT).onUp.add(function(){gameStream.emit('Movement',  {x: jimmy.x, y: jimmy.y, dir:"right", name: jimmy.name, key: false})});
+        game.input.keyboard.addKey(Phaser.Keyboard.UP).onUp.add(function(){gameStream.emit('Movement',  {x: jimmy.x, y: jimmy.y, dir:"up", name: jimmy.name, key: false})});
+        game.input.keyboard.addKey(Phaser.Keyboard.DOWN).onUp.add(function(){gameStream.emit('Movement',  {x: jimmy.x, y: jimmy.y, dir:"down", name: jimmy.name, key: false})});
       
         game.input.keyboard.addKey(Phaser.Keyboard.SPACEBAR).onDown.add(function(){
         if (jimmy.alive){
-        gameStream.emit('bombSet',{x: jimmy.x, y: jimmy.y,type:"bomb"});
+        gameStream.emit('bombSet',{x: jimmy.x, y: jimmy.y, type:"bomb"});
         setbomb(jimmy, 'bomb');
           }
         });
@@ -56,14 +56,14 @@ function create() {
         // checkposition.timer.start();
 
 
-        game.input.keyboard.addKey(Phaser.Keyboard.A).onDown.add(function(){gameStream.emit('glennLeft',  {key: true})});
-        game.input.keyboard.addKey(Phaser.Keyboard.D).onDown.add(function(){gameStream.emit('glennRight',  {key: true})});
-        game.input.keyboard.addKey(Phaser.Keyboard.W).onDown.add(function(){gameStream.emit('glennUp',  {key: true})});
-        game.input.keyboard.addKey(Phaser.Keyboard.S).onDown.add(function(){gameStream.emit('glennDown',  {key: true})});
-        game.input.keyboard.addKey(Phaser.Keyboard.A).onUp.add(function(){gameStream.emit('glennLeft',  {x: glenn.x, y: glenn.y, key: false})});
-        game.input.keyboard.addKey(Phaser.Keyboard.D).onUp.add(function(){gameStream.emit('glennRight',  {x: glenn.x, y: glenn.y, key: false})});
-        game.input.keyboard.addKey(Phaser.Keyboard.W).onUp.add(function(){gameStream.emit('glennUp',  {x: glenn.x, y: glenn.y, key: false})});
-        game.input.keyboard.addKey(Phaser.Keyboard.S).onUp.add(function(){gameStream.emit('glennDown',  {x: glenn.x, y: glenn.y, key: false})});
+        game.input.keyboard.addKey(Phaser.Keyboard.A).onDown.add(function(){gameStream.emit('Movement',  {dir:"left", name: glenn.name, key: true})});
+        game.input.keyboard.addKey(Phaser.Keyboard.D).onDown.add(function(){gameStream.emit('Movement',  {dir:"right", name: glenn.name, key: true})});
+        game.input.keyboard.addKey(Phaser.Keyboard.W).onDown.add(function(){gameStream.emit('Movement',  {dir:"up", name: glenn.name, key: true})});
+        game.input.keyboard.addKey(Phaser.Keyboard.S).onDown.add(function(){gameStream.emit('Movement',  {dir:"down", name: glenn.name, key: true})});
+        game.input.keyboard.addKey(Phaser.Keyboard.A).onUp.add(function(){gameStream.emit('Movement',  {x: glenn.x, y: glenn.y, dir:"left", name: glenn.name, key: false})});
+        game.input.keyboard.addKey(Phaser.Keyboard.D).onUp.add(function(){gameStream.emit('Movement',  {x: glenn.x, y: glenn.y, dir:"right", name: glenn.name, key: false})});
+        game.input.keyboard.addKey(Phaser.Keyboard.W).onUp.add(function(){gameStream.emit('Movement',  {x: glenn.x, y: glenn.y, dir:"up", name: glenn.name, key: false})});
+        game.input.keyboard.addKey(Phaser.Keyboard.S).onUp.add(function(){gameStream.emit('Movement',  {x: glenn.x, y: glenn.y, dir:"down", name: glenn.name, key: false})});
       
         game.input.keyboard.addKey(Phaser.Keyboard.Z).onDown.add(function(){
         if (glenn.alive){
@@ -97,17 +97,21 @@ function create() {
     tombs.physicsBodyType = Phaser.Physics.ARCADE;
 
 
-    jimmy = game.add.sprite(50, 100, 'jimmy');
+    jimmy = game.add.sprite(660, 100, 'jimmy');
+    jimmy.name = "jimmy";
     jimmy.scale.set(0.14);
     game.physics.enable(jimmy, Phaser.Physics.ARCADE);
     jimmy.enableBody = true;
     jimmy.body.collideWorldBounds = true;
+    jimmy.body.mass = 0.1;
 
-    glenn = game.add.sprite(660, 100, 'glenn');
+    glenn = game.add.sprite(50, 100, 'glenn');
+    glenn.name = "glenn";
     glenn.scale.set(0.1);
     game.physics.enable(glenn, Phaser.Physics.ARCADE);
     glenn.enableBody = true;
     glenn.body.collideWorldBounds = true;
+    glenn.body.mass = 0.1;
 
 
     for (var j = 0; j < 5; j++)
@@ -121,83 +125,81 @@ function create() {
     }
 }
 
-  gameStream.on('jimmyLeft', function(data){
+  gameStream.on('Movement', function(data){
+    if (data.name == "jimmy") {
+    if (data.dir == "left") {
     jimmyLeft = data.key
-       if (data.x) {
+  }
+  else if (data.dir == "right"){
+    jimmyRight = data.key
+  }
+  else if (data.dir == "up"){
+    jimmyUp = data.key
+  }
+  else if (data.dir == "down"){
+    jimmyDown = data.key
+  };
+
+  if (!data.key) {
       jimmy.x = data.x;
       jimmy.y = data.y;
     };
-  })
 
-  gameStream.on('jimmyRight', function(data){
-   jimmyRight = data.key
-  
-    if (data.x) {
-      jimmy.x = data.x;
-      jimmy.y = data.y;
-    };
-  })
-
-  gameStream.on('jimmyDown', function(data){
-   jimmyDown = data.key
-      if (data.x) {
-      jimmy.x = data.x;
-      jimmy.y = data.y;
-    };
-  })
-
-  gameStream.on('jimmyUp', function(data){
-   jimmyUp = data.key
-      if (data.x) {
-      jimmy.x = data.x;
-      jimmy.y = data.y;
-    };
-  })
-
-  gameStream.on('glennLeft', function(data){
+  }
+  else if (data.name == "glenn") {
+    if (data.dir == "left") {
     glennLeft = data.key
-       if (data.x) {
+  }
+  else if (data.dir == "right"){
+    glennRight = data.key
+  }
+  else if (data.dir == "up"){
+    glennUp = data.key
+  }
+  else if (data.dir == "down"){
+    glennDown = data.key
+  };
+
+  if (!data.key) {
       glenn.x = data.x;
       glenn.y = data.y;
     };
+
+  }
   })
 
-  gameStream.on('glennRight', function(data){
-   glennRight = data.key
-  
-    if (data.x) {
-      glenn.x = data.x;
-      glenn.y = data.y;
+
+
+    gameStream.on('playerDie', function(data){
+      if (data.name == "jimmy" && jimmy.alive) {
+        var tomb = tombs.create(data.x, data.y+40, 'tomb');
+        jimmy.kill()
+    }
+    else if (data.name == "glenn" && glenn.alive) {
+      var tomb = tombs.create(data.x, data.y+40, 'tomb');
+      glenn.kill()
     };
   })
 
-  gameStream.on('glennDown', function(data){
-   glennDown = data.key
-      if (data.x) {
-      glenn.x = data.x;
-      glenn.y = data.y;
-    };
-  })
+  // gameStream.on('jimmyMove', function(data){
+  //   jimmy.x = data.x;
+  //   jimmy.y = data.y;
+  // })
 
-  gameStream.on('glennUp', function(data){
-   glennUp = data.key
-      if (data.x) {
-      glenn.x = data.x;
-      glenn.y = data.y;
-    };
-  })
-
-  gameStream.on('jimmyMove', function(data){
-    jimmy.x = data.x;
-    jimmy.y = data.y;
-  })
-
-  gameStream.on('glennMove', function(data){
-    glenn.x = data.x;
-    glenn.y = data.y;
-  })
+  // gameStream.on('glennMove', function(data){
+  //   glenn.x = data.x;
+  //   glenn.y = data.y;
+  // })
 
   gameStream.on('bombSet', function(data){
+    if (data.type=="star"){
+      glenn.x = data.x;
+      glenn.y = data.y;
+    }
+    else{
+      jimmy.x = data.x;
+      jimmy.y = data.y;
+    }
     setbomb(data, data.type);
   })
 
@@ -208,6 +210,7 @@ function update(){
   game.physics.arcade.collide(jimmy, bombs);
   game.physics.arcade.collide(jimmy, fires, die, null, this);
   game.physics.arcade.collide(boxes, fires, killfire, null, this);
+  game.physics.arcade.collide(bombs, fires, explode, null, this);
 
   game.physics.arcade.collide(glenn, boxes);
   game.physics.arcade.collide(glenn, bombs);
@@ -281,6 +284,7 @@ function setbomb(player, type){
   
       self.bomb = bombs.create(player.x, player.y+45, type);
       bomb.body.collideWorldBounds = true;
+      bomb.body.mass = 0.01;
       if(type=="star"){
         self.bomb.scale.set(0.13);
       }
@@ -293,18 +297,21 @@ function setbomb(player, type){
 
 
 function explode(bom){
-    var fire1 = fires.create(bom.x, bom.y, 'fire');
-    fire1.body.velocity.y = -300;
-    var fire2 = fires.create(bom.x, bom.y, 'fire');
-    fire2.body.velocity.y = 300;
-    var fire3 = fires.create(bom.x, bom.y, 'fire');
-    fire3.body.velocity.x = -300;
-    var fire4 = fires.create(bom.x, bom.y, 'fire');
-    fire4.body.velocity.x = 300;
-    fire1.scale.set(0.7);
-    fire2.scale.set(0.7);
-    fire3.scale.set(0.7);
-    fire4.scale.set(0.7);
+    if(bom.alive)
+    {
+      var fire1 = fires.create(bom.x, bom.y, 'fire');
+      fire1.body.velocity.y = -300;
+      var fire2 = fires.create(bom.x, bom.y, 'fire');
+      fire2.body.velocity.y = 300;
+      var fire3 = fires.create(bom.x, bom.y, 'fire');
+      fire3.body.velocity.x = -300;
+      var fire4 = fires.create(bom.x, bom.y, 'fire');
+      fire4.body.velocity.x = 300;
+      fire1.scale.set(0.7);
+      fire2.scale.set(0.7);
+      fire3.scale.set(0.7);
+      fire4.scale.set(0.7);
+  }
     bom.kill();
     console.log("boom!")
 
@@ -316,8 +323,13 @@ function killfire(boxes, fire){
 }
 
 function die(player, fire){
-    var tomb = tombs.create(player.x, player.y+40, 'tomb');
-    player.kill();
+  gameStream.emit('playerDie',  {x: player.x, y: player.y, name: player.name})
+
+    if (player.alive) {
+      var tomb = tombs.create(player.x, player.y+40, 'tomb');
+      player.kill();
+    };
+
 }
